Dev: watchdog: split class Watchdog into watchdog.py from bootstrap.py
